Implement basic eager packet path
This adds changes so that an incoming eager packet will land in an
eager buffer and generate an interrupt.  Also includes misc changes
implemented on the way.

Details:
o Change dd->kregbase and dd->kregend to u8* so magic pointer
  arithmetic does not occur in ui_lseek().
o Because of the above change, fix up pointer arithmetic in
  qib_remap_ioaddr32().
o Because of the above change, #if 0 out init_chip_wc_pat()
  (currently not called) until we're ready to implement it.
o Remove unused variable kreg_base in qib_diag_write().
o Uncomment write_uctxt_csr() - it is now used.
o Enable the port in bringup_serdes().  The function is otherwise
  still not implemented.
o The physical address stored in RcvArray.RT_Addr needs to be
  shifted to a 4KB boundary.
o Print which argument as a string for {get,set}_ib_cfg().
o In one_rcvctrl(), only do extra enable actions when
  RcvCtxtCtrl.Enable transitions from 0 to 1.
o In one_rcvctrl(), after RcvCtxtCtrl.Enable transitions from 0 to 1,
  Set RcvHdrHead.Counter to a non-zero value so we receive interrupts
  when packets arrive.
o Have iblink_state() always report IB_PORT_ACTIVE.
o Have ibphys_portstate() always report IB_PHYSPORTSTATE_LINKUP.
o Have ib_updown() print more information.
o Add more hand-defined CSRs.

+++ hfi.h
@@ -717,9 +717,9 @@ struct qib_devdata {
 	struct device *ui_device;
 
 	/* mem-mapped pointer to base of chip regs */
-	u64 __iomem *kregbase;
+	u8 __iomem *kregbase;
 	/* end of mem-mapped chip space excluding sendbuf and user regs */
-	u64 __iomem *kregend;
+	u8 __iomem *kregend;
 	/* physical address of chip for io_remap, etc. */
 	resource_size_t physaddr;
 	/* qib_cfgctxts pointers */
